+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-08-29
+
 ### Added
 
 - Add compatibility with Seasons.
@@ -52,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version.
 
-[unreleased]: https://github.com/nbusseneau/InstantFertilizer/compare/0.3.0...HEAD
+[unreleased]: https://github.com/nbusseneau/InstantFertilizer/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/nbusseneau/InstantFertilizer/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/nbusseneau/InstantFertilizer/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/nbusseneau/InstantFertilizer/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/nbusseneau/InstantFertilizer/compare/0.1.0...0.1.1